requirements.txt
fastapi
uvicorn[standard]

sqlalchemy>=2.0
asyncpg
psycopg[binary]
alembic

python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.1

pydantic-settings
